% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanpy.R
\name{anndataR}
\alias{anndataR}
\title{basic interface to anndata}
\usage{
anndataR()
}
\value{
basiliskRun result with import from reticulate, typically a Module
}
\description{
basic interface to anndata
}
\examples{
ad <- anndataR()
ad
ad$read
}
