% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilisk.R
\docType{data}
\name{bsklenv}
\alias{bsklenv}
\title{python declarations}
\format{
An object of class \code{BasiliskEnvironment} of length 1.
}
\usage{
bsklenv
}
\description{
python declarations
}
\keyword{datasets}
