% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_citeseq_pbmcs.R
\name{cache_citeseq_5k10k_pbmcs}
\alias{cache_citeseq_5k10k_pbmcs}
\title{grab scvi-tools-processed PBMC CITE-seq data in anndata format (gzipped) from Open Storage Network}
\usage{
cache_citeseq_5k10k_pbmcs()
}
\value{
invisibly, the path to the .h5ad file
}
\description{
grab scvi-tools-processed PBMC CITE-seq data in anndata format (gzipped) from Open Storage Network
}
\note{
Original h5ad files obtained using scvi-tools 0.18.0 scvi.data.pbmcs_10x_cite_seq,
then processed according to steps in the scviR vignette, which follow the
[scvi-tools tutorial](https://colab.research.google.com/github/scverse/scvi-tutorials/blob/0.18.0/totalVI.ipynb) by Gayoso et al.
}
\examples{
h5path = cache_citeseq_5k10k_pbmcs()
cmeta = rhdf5::h5ls(h5path)
dim(cmeta)
head(cmeta, 17)
}
