% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_citeseq_pbmcs.R
\name{cache_citeseq_5k10k_tutvae}
\alias{cache_citeseq_5k10k_tutvae}
\title{grab scvi-tools VAE instance built on the PBMC datasets following the tutorial}
\usage{
cache_citeseq_5k10k_tutvae()
}
\value{
invisibly, the path to the .zip file holding the fitted VAE and associated data
}
\description{
grab scvi-tools VAE instance built on the PBMC datasets following the tutorial
}
\note{
VAE construction followed tutorial at 
`https://docs.scvi-tools.org/en/stable/tutorials/notebooks/totalVI.html`.
}
\examples{
zpath = cache_citeseq_5k10k_tutvae()
td = tempdir()
utils::unzip(zpath, exdir=td)
vaedir = paste0(td, "/vae2_ov")
scvi = scviR()
adm = anndataR()
hpath = cache_citeseq_5k10k_pbmcs()
adata = adm$read(hpath)
mod = scvi$model$`_totalvi`$TOTALVI$load(vaedir, adata, use_gpu=FALSE)
mod
}
