% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_citeseq_pbmcs.R
\name{cacheCiteseq5k10kPbmcs}
\alias{cacheCiteseq5k10kPbmcs}
\title{grab scvi-tools-processed PBMC CITE-seq data in anndata format (gzipped) from Open Storage Network}
\usage{
cacheCiteseq5k10kPbmcs()
}
\value{
invisibly, the path to the .h5ad file
}
\description{
grab scvi-tools-processed PBMC CITE-seq data in anndata format (gzipped) from Open Storage Network
}
\note{
Original h5ad files obtained using scvi-tools 0.18.0 scvi.data.pbmcs_10x_cite_seq,
then processed according to steps in the scviR vignette, which follow the
[scvi-tools tutorial](https://colab.research.google.com/github/scverse/scvi-tutorials/blob/0.18.0/totalVI.ipynb) by Gayoso et al.

It may be advantageous to set `options(timeout=3600)` or to allow an even greater
time for internet downloads, if working at a relatively slow network connection.
}
\examples{
h5path = cacheCiteseq5k10kPbmcs()
cmeta = rhdf5::h5ls(h5path)
dim(cmeta)
head(cmeta, 17)
}
