% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_citeseq.R
\name{cacheCiteseqHDPdata}
\alias{cacheCiteseqHDPdata}
\title{retrieve and cache a 349-protein CITE-seq dataset as employed in
scvi-tools tutorial}
\usage{
cacheCiteseqHDPdata()
}
\description{
retrieve and cache a 349-protein CITE-seq dataset as employed in
scvi-tools tutorial
}
