% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clusters.adt}
\alias{clusters.adt}
\title{ADT-based cluster labels for 7472 cells in OSCA chapter 12 analysis}
\format{
factor
}
\usage{
clusters.adt
}
\description{
ADT-based cluster labels for 7472 cells in OSCA chapter 12 analysis
}
\keyword{datasets}
