% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clusters.rna}
\alias{clusters.rna}
\title{mRNA-based cluster labels for 7472 cells in OSCA chapter 12 analysis}
\format{
factor
}
\usage{
data(clusters.rna)
}
\description{
mRNA-based cluster labels for 7472 cells in OSCA chapter 12 analysis
}
\keyword{datasets}
