% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citeseqApp.R
\name{explore_subcl}
\alias{explore_subcl}
\title{app to explore diversity in RNA-subclusters within ADT clusters}
\usage{
explore_subcl(sce, inlist, adtcls)
}
\arguments{
\item{sce}{a SingleCellExperiment with altExp with ADT quantification}

\item{inlist}{list of SingleCellExperiments (SCEs) formed by scran::quickSubCluster}

\item{adtcls}{vector of ADT cluster assignments}
}
\description{
app to explore diversity in RNA-subclusters within ADT clusters
}
\note{
TSNE should already be available in `altExp(sce)`; follow OSCA book 12.5.2.  If using
example, set `ask=FALSE`.
}
\examples{
if (interactive()) {
 data(sce)
 data(all.sce)
 data(clusters.adt)
 explore_subcl( sce, all.sce, clusters.adt )
}
}
