% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_citeseq_pbmcs.R
\name{get_citeseq_pbmcs}
\alias{get_citeseq_pbmcs}
\title{helper to get the processed anndata for CITE-seq PBMCs from scvi-tools tutorial}
\usage{
get_citeseq_pbmcs()
}
\description{
helper to get the processed anndata for CITE-seq PBMCs from scvi-tools tutorial
}
\examples{
get_citeseq_pbmcs()
}
