% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_citeseq_pbmcs.R
\name{get_citeseq_tutvae}
\alias{get_citeseq_tutvae}
\title{helper to get the tutorial VAE for PBMCs from scvi-tools tutorial}
\usage{
get_citeseq_tutvae()
}
\description{
helper to get the tutorial VAE for PBMCs from scvi-tools tutorial
}
\examples{
get_citeseq_tutvae()
}
