% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_citeseq_pbmcs.R
\name{get_pro_5k10k_adata}
\alias{get_pro_5k10k_adata}
\title{get an anndata reference to 5k10k protein after totalVI from tutorial}
\usage{
get_pro_5k10k_adata()
}
\description{
get an anndata reference to 5k10k protein after totalVI from tutorial
}
\examples{
get_pro_5k10k_adata()
}
