% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citeseqApp.R
\name{get_subcl_LM}
\alias{get_subcl_LM}
\title{get lmFit for heterogeneity across subclusters}
\usage{
get_subcl_LM(inlist, clname)
}
\arguments{
\item{inlist}{list of SingleCellExperiments (SCEs) formed by scran::quickSubCluster}

\item{clname}{character(1) name of cluster SCE to assess}
}
\description{
get lmFit for heterogeneity across subclusters
}
\note{
It is assumed that 'logcounts' is an assay element,
and that 'subcluster' is a colData element of each SCE in inlist
}
\examples{
data(all.sce)
lm3 = get_subcl_LM(all.sce, "3")
names(lm3)
}
