% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_citeseq_pbmcs.R
\name{get_totalVI_5k10k_adata}
\alias{get_totalVI_5k10k_adata}
\title{get anndata reference to full totalVI processing of 5k10k data}
\usage{
get_totalVI_5k10k_adata()
}
\value{
python reference to anndata
}
\description{
get anndata reference to full totalVI processing of 5k10k data
}
\examples{
full = get_totalVI_5k10k_adata()
full
}
