% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_citeseq_pbmcs.R
\name{getCh12AllSce}
\alias{getCh12AllSce}
\title{get list of cluster-specific SCE for 10k PBMC annotated as in OSCA book chapter 12}
\usage{
getCh12AllSce()
}
\value{
SimpleList of SingleCellExperiment instances
}
\description{
get list of cluster-specific SCE for 10k PBMC annotated as in OSCA book chapter 12
}
\note{
This is a list of SingleCellExperiment instances with data on a total of 7472 cells from a 10x
CITE-seq experiment.  An altExp component in each list element includes
antibody-derived tag (ADT) counts on 17 proteins.  The data are acquired and
processed as described in ch 12 of the OSCA book, circa February 2023.
List elements correspond to mRNA-based sub-clusters of ADT-based clusters.
}
\examples{
ch12_allsce = getCh12AllSce()
vapply(ch12_allsce, ncol, numeric(1))
}
