% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_citeseq_pbmcs.R
\name{getCh12Sce}
\alias{getCh12Sce}
\title{get SCE for 10k PBMC annotated as in OSCA book chapter 12}
\usage{
getCh12Sce(clear_cache = FALSE)
}
\arguments{
\item{clear_cache}{logical(1) will delete relevant entries in available cache before continuing, defaults to FALSE}
}
\value{
SingleCellExperiment instance
}
\description{
get SCE for 10k PBMC annotated as in OSCA book chapter 12
}
\note{
This is a SingleCellExperiment instance with data on 7472 cells from a 10x
CITE-seq experiment.  An altExp component includes
antibody-derived tag (ADT) counts on 17 proteins.  The data are acquired and
processed as described in ch 12 of the OSCA book, circa February 2023.
A metadata element (se.averaged) includes the result of averaging protein abundance
estimates within ADT-based clusters, as is done to give rise to Figure 12.8 of
the OSCA book.
}
\examples{
ch12sce <- getCh12Sce()
ch12sce
}
