% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_citeseq_pbmcs.R
\name{getCiteseq5k10kPbmcs}
\alias{getCiteseq5k10kPbmcs}
\title{helper to get the processed anndata for CITE-seq PBMCs from scvi-tools tutorial}
\usage{
getCiteseq5k10kPbmcs()
}
\value{
python reference to anndata
}
\description{
helper to get the processed anndata for CITE-seq PBMCs from scvi-tools tutorial
}
\note{
It may be advantageous to set `options(timeout=3600)` or to allow an even greater
time for internet downloads, if working at a relatively slow network connection.
}
\examples{
getCiteseq5k10kPbmcs()
}
