% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_citeseq_pbmcs.R
\name{getCiteseqTutvae}
\alias{getCiteseqTutvae}
\title{helper to get the tutorial VAE for PBMCs from scvi-tools tutorial}
\usage{
getCiteseqTutvae(use_gpu = FALSE)
}
\arguments{
\item{use_gpu}{logical(1), defaulting to FALSE, passed to TOTALVI.load}
}
\value{
python reference to anndata
}
\description{
helper to get the tutorial VAE for PBMCs from scvi-tools tutorial
}
\examples{
getCiteseqTutvae()
}
