% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_citeseq_pbmcs.R
\name{getPro5k10kAdata}
\alias{getPro5k10kAdata}
\title{get an anndata reference to 5k10k protein after totalVI from tutorial}
\usage{
getPro5k10kAdata()
}
\value{
python reference to anndata
}
\description{
get an anndata reference to 5k10k protein after totalVI from tutorial
}
\note{
It may be advantageous to set `options(timeout=3600)` or to allow an even greater
time for internet downloads, if working at a relatively slow network connection.
}
\examples{
getPro5k10kAdata()
}
