% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_citeseq_pbmcs.R
\name{getTotalVI5k10kAdata}
\alias{getTotalVI5k10kAdata}
\title{get anndata reference to full totalVI processing of 5k10k data}
\usage{
getTotalVI5k10kAdata()
}
\value{
python reference to anndata
}
\description{
get anndata reference to full totalVI processing of 5k10k data
}
\examples{
full = getTotalVI5k10kAdata()
full
}
