% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanpy.R
\name{MuDataR}
\alias{MuDataR}
\title{basic interface to MuData}
\usage{
MuDataR()
}
\value{
basiliskRun result with import from reticulate, typically a Module
}
\description{
basic interface to MuData
}
\examples{
md <- MuDataR()
md
head(names(md))
}
