% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanpy.R
\name{muonR}
\alias{muonR}
\title{basic interface to muon}
\usage{
muonR()
}
\value{
basiliskRun result with import from reticulate, typically a Module
}
\description{
basic interface to muon
}
\examples{
md <- muonR()
md
head(names(md))
}
