% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scverse_helper.R
\name{py_help2}
\alias{py_help2}
\title{helper to get text from python help utility -- may need handling through basilisk}
\usage{
py_help2(object)
}
\arguments{
\item{object}{a reference to a python module typically with class 'python.builtin.module'}
}
\description{
helper to get text from python help utility -- may need handling through basilisk
}
