% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scverse_helper.R
\name{pyHelp2}
\alias{pyHelp2}
\title{helper to get text from python help utility -- may need handling through basilisk}
\usage{
pyHelp2(object)
}
\arguments{
\item{object}{a reference to a python module typically with class 'python.builtin.module'}
}
\value{
character vector of lines from python help result
}
\description{
helper to get text from python help utility -- may need handling through basilisk
}
