% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scverse_helper.R
\name{scanpyHelper}
\alias{scanpyHelper}
\title{shiny app that helps access documentation on python-accessible components}
\usage{
scanpyHelper()
}
\value{
shinyApp instance
}
\description{
shiny app that helps access documentation on python-accessible components
}
