% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanpy.R
\name{scanpyR}
\alias{scanpyR}
\title{basic interface}
\usage{
scanpyR()
}
\value{
basiliskRun result with import from reticulate, typically a Module
}
\description{
basic interface
}
\examples{
sc = scanpyR()
sc
sc$pp
}
