% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scverse_helper.R
\name{scverse_helper}
\alias{scverse_helper}
\title{shiny app that helps access documentation on python-accessible components}
\usage{
scverse_helper()
}
\description{
shiny app that helps access documentation on python-accessible components
}
