% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scverse_helper.R
\name{scverseHelper}
\alias{scverseHelper}
\title{shiny app that helps access documentation on python-accessible components}
\usage{
scverseHelper()
}
\value{
shinyApp instance
}
\description{
shiny app that helps access documentation on python-accessible components
}
