% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scvi.R
\name{scviR}
\alias{scviR}
\title{basic interface}
\usage{
scviR()
}
\value{
basiliskRun result with import from reticulate, typically a Module
}
\description{
basic interface
}
\examples{
scvi <- scviR()
scvi
scvi$model
}
