% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_citeseq_pbmcs.R
\name{cacheCiteseq5k10kTutvae}
\alias{cacheCiteseq5k10kTutvae}
\title{grab scvi-tools VAE instance built on the PBMC datasets following the tutorial}
\usage{
cacheCiteseq5k10kTutvae()
}
\value{
invisibly, the path to the .zip file holding the fitted VAE and associated data
}
\description{
grab scvi-tools VAE instance built on the PBMC datasets following the tutorial
}
\note{
VAE construction followed tutorial at
`https://docs.scvi-tools.org/en/stable/tutorials/notebooks/totalVI.html`.

It may be advantageous to set `options(timeout=3600)` or to allow an even greater
time for internet downloads, if working at a relatively slow network connection.
}
\examples{
zpath <- cacheCiteseq5k10kTutvae()
td <- tempdir()
utils::unzip(zpath, exdir = td)
vaedir <- paste0(td, "/vae2_ov")
scvi <- scviR()
adm <- anndataR()
hpath <- cacheCiteseq5k10kPbmcs()
adata <- adm$read(hpath)
mod <- scvi$model$`_totalvi`$TOTALVI$load(vaedir, adata) #, use_gpu = FALSE)
mod
}
